+++ demo/GodelTech.Microservices.Security.Demo.Mvc/Controllers/ApiRequestController.cs
@@ -13,6 +13,15 @@ namespace GodelTech.Microservices.Security.Demo.Mvc.Controllers
             _httpClientFactory = httpClientFactory;
         }
 
+        public async Task<IActionResult> CallApiAsUser()
+        {
+            var client = _httpClientFactory.CreateClient("UserClient");
+
+            ViewBag.Response = await client.GetStringAsync("fakes/1");
+
+            return View("CallApi");
+        }
+
         public async Task<IActionResult> CallApiAsClient()
         {
             var client = _httpClientFactory.CreateClient("ApiClient");
